Replace RemoveOnConditions with KillsSelf.
Upgrades offer more flexibility than prerequisites.
